feat: 장바구니 아이템 타입 분리 및 코드 정리
- CartItem 인터페이스를 새로운 index.ts 파일로 분리하여 코드의 일관성을 높였습니다.
- cartCalculator.ts에서 불필요한 CartItem 정의를 제거하여 코드의 간결성을 향상시켰습니다.
