+++ src/advanced/features/cart/types/index.ts
@@ -1,0 +1,12 @@
+/**
+ * 장바구니 아이템 타입
+ */
+export interface CartItem {
+  id: string;
+  name: string;
+  val: number;
+  originalVal: number;
+  quantity: number;
+  onSale: boolean;
+  suggestSale: boolean;
+}
